app/requirements.txt
streamlit 
streamlit_option_menu
pandas
numpy
matplotlib.pyplot
seaborn
oracledb
geopy.geocoders
folium
streamlit_folium
requests
os
supabase
langchain.embeddings
langchain.vectorstores
langchain.llms
langchain.chains
langchain_core.runnables
langchain.prompts
dotenv

